requirements.txt
celery==5.3.6
Django==5.0.3
djangorestframework-simplejwt
django-celery-results
psycopg2-binary
selenium
bs4
Redis
flower==1.2.0
